move lame to module
