+++ modules/fun/index.ts
