+++ modules/fun/mentions/magic8.ts
